requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
attrs==23.2.0
beautifulsoup4==4.12.3
Brotli 
certifi 
charset-normalizer 
contourpy==1.2.1
cycler==0.12.1
docopt==0.6.2
dominate==2.9.1
einops==0.7.0
filelock 
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
gdown==5.1.0
huggingface-hub==0.22.2
hydra-core==1.3.2
idna 
imageio==2.34.0
imgaug==0.4.0
Jinja2 
joblib==1.4.0
kiwisolver==1.4.5
lazy_loader==0.4
lightning==2.3.0
lightning-utilities==0.11.3.post0
MarkupSafe 
matplotlib==3.9.0
# mkl-fft 
# mkl-random 
# mkl-service==2.4.0
mpmath 
multidict==6.0.5
networkx 
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
pillow==10.3.0
progressbar==2.5
pyparsing==3.1.2
PySocks 
python-dateutil==2.9.0.post0
pytorch-lightning==2.3.0
pytz==2024.1
PyYAML 
requests 
safetensors==0.4.3
scikit-image==0.23.1
scikit-learn==1.4.2
scipy==1.13.0
setuptools==68.2.2
shapely==2.0.4
six==1.16.0
soupsieve==2.5
sympy 
threadpoolctl==3.4.0
tifffile==2024.4.18
timm==0.9.16
torch==2.2.2
torchaudio==2.2.2
torchmetrics==1.4.0.post0
torchvision==0.17.2
tqdm==4.66.2
typing_extensions 
tzdata==2024.1
urllib3 
wheel==0.41.2
yarl==1.9.4
